+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Supported binding parameters for `database/sql` driver
-* Supported `ydb.WithTablePathPrefix(tablePathPrefix)` connector option
+* Added packaged `bind` with bindings options `WithTablePathPrefix(tablePathPrefix)` and `Params()`
+* Added `ydb.WithBindings` connector option
 * Fixed bug with ignoring default scan and data query options
 * Fixed topic retry policy callback call: not call it with nil error 
 * Deprecated `trace.Driver.OnConnTake` event callback